+++ src/cli.ts
@@ -3,6 +3,7 @@ import path from "node:path/posix";
 import { build as viteBuild, createLogger } from "vite";
 import colors from "picocolors";
 import * as babel from "@babel/core";
+import { displayTime, prettySize } from "./utils";
 
 const cjsSrcFile = "temp/index.cjs.js";
 const esmSrcFile = "temp/index.es.js";
@@ -10,42 +11,6 @@ const dstDir = "dist";
 const cjsDstFile = `${dstDir}/cjs/index.cjs.js`;
 const esmDstFile = `${dstDir}/esm/index.esm.js`;
 const logger = createLogger();
-
-function displayTime(time: number): string {
-    // display: {X}ms
-    if (time < 1000) {
-        return `${time}ms`;
-    }
-
-    time = time / 1000;
-
-    // display: {X}s
-    if (time < 60) {
-        return `${time.toFixed(2)}s`;
-    }
-
-    const mins = parseInt((time / 60).toString(), 10);
-    const seconds = time % 60;
-    const minuteString = `${mins}m`;
-    const secondString = seconds < 1 ? "" : ` ${seconds.toFixed(0)}s`;
-
-    // display: {X}m {Y}s
-    return `${minuteString}m${secondString}`;
-}
-
-function prettySize(size: number): string {
-    if (size < 1024) {
-        return `${size} B`;
-    } else if (size < 1024 * 1024) {
-        const divisor = 1024;
-        const rounded = (size / divisor).toFixed(2);
-        return `${rounded} kB`;
-    } else {
-        const divisor = 1024 * 1024;
-        const rounded = (size / divisor).toFixed(2);
-        return `${rounded} mB`;
-    }
-}
 
 async function transpile(src: string, dst: string): Promise<void> {
     const dstMap = `${dst}.map`;
