+++ src/utils/index.ts
@@ -1,6 +1,8 @@
 export { detectInternalDependencies } from "./detect-internal-dependencies";
 export { detectMonorepoPackages } from "./detect-monorepo-packages";
 export { detectVueMajor } from "./detect-vue-major";
+export { displayTime } from "./display-time";
 export { extractAugmentations } from "./extract-augmentations";
 export { prettyList } from "./pretty-list";
+export { prettySize } from "./pretty-size";
 export { readJsonFile } from "./read-json-file";
